# web/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Stores and Retrieves information needed"
  version: "1.0.0"
  title: "Team 4 Storage API"
host: "ec2-54-69-164-246.us-west-2.compute.amazonaws.com:8000"
basePath: "/v1"
schemes:
- "http"
paths:
...
  /primary/ver={version}/:
    post:
      tags:
      - "primary"
      summary: "Update the existing primary"
      description: ""
      operationId: "post_primary"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "version"
        in: "path"
        description: "The version of the problem being manipulated"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "problem"
        description: "Problem object that needs to be updated."
        required: true
        schema:
          $ref: "#/definitions/Body"
      responses:
        200:
          description: "The new version number for the problem"
          schema:
            type: "integer"
            format: "int32"
        404:
          description: "Problem not found"
        412:
          description: "Incorrect version number"
          schema:
            type: "integer"
            format: "int32"
            description: "latest version number"
        500:
          description: "Incorrect JSON"
      x-swagger-router-controller: "swagger_server.controllers.primary_controller"
...
definitions:
  Body:
    type: "object"
...